snapalloc: Add snapalloc implementation
This change implements the ISnapAlloc and ISnapMapper interfaces.

ISnapAlloc allocates buffer memory and
ISnapMapper is used for mapping and accessing the buffer
in the current process' memory space.

For ISnapAlloc, the allocation
requirements are queried from SnapConstraintProviders
and combined by the SnapConstraintManager.
SnapMemAllocator does the memory allocation operations.

For ISnapMapper, SnapMemAllocator also handles memory-related operations.
SnapMetadataManager is used to read or write metadata
(information about the buffer other than the buffer's contents -
e.g., width and height of the buffer).

Change-Id: If51eec33243c174b697a41a9d8d72ea9e5ebda74
CRs-Fixed: 3651916