sdm: return no error if the driver returns empty edid blob
During driver initialization if the sink returns invalid/corrupted edid,
the connector cache would contain empty edid blob. However, the driver
has a failsafe in place where the default 640x480 mode would be set
during the get modes. Currently SDM is returning error code upstream if
edid blob is empty which would not enable the display. As per the spec,
default mode should be set in case of edid failure and this test is
failing during link CTS. This change would return no error in this
case and print a warning so that this failsafe can kick in.

Change-Id: Idf4ad4b4af81c7da3c1d03e545fa5300eeb26812
Signed-off-by: Sandeep Gangadharaiah <quic_sandgang@quicinc.com>