sdmclient: Reject same CWB buffer, if already exists in queue
Multiple instances of CWB request with same buffer id can't be queued
in CWB manager. Because it may cause race condition, that one is trying
to write the buffer and another trying to read at a time, which may
cause driver commit failure. So, just need to reject CWB request with
buffer, which is already available for CWB processing in a queue.

Change-Id: Iab65205bfa30d1f7724f5096f6150f3b4659ea6d