display-core: update handle id for demura layer
All plane properties get updated to driver if there is any change
to frame buffer object. Null handle id is detected as frame buffer
object change. Handle id is not updated for demura and ABC layers
and this is leading to updating all plane properties for all commit
cycles. This change updates handle id for demura and ABC layers.

Change-Id: I6c055b5bfddea6f563355d6fd63b9f4545c09bdb