display-core: Add check if GPU HW is available or not, for qultivate device
For GPU faulty device, During video playback, RC which is on top of
all layers do not get alpha values. Due to this even if display is
functional, video playback will be blank.

To avoid black screen, we are checking if GPU HW is available or not
and informing SF that Display Decoration is not supported.

Change-Id: I2da764e43cb639cca2e790c12d48ad3c4fe9bfe1